+++ examples/dogsvscats/train_test.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 	"testing"
 
-	_ "github.com/gomlx/gomlx/backends/xla"
+	_ "github.com/gomlx/gomlx/backends/default"
 )
 
 var (